main
The Smart Handwash Dispenser is a touchless system designed to improve hygiene by automating the soap dispensing process. The goal of this project was to create a user-friendly, efficient, and cost-effective way to dispense liquid soap without the need for physical contact, which is especially useful in public restrooms and healthcare environments.